"discover_inductive" instead of "discovery_inductive"
In version 2.0.0 of  `pm4py`, the function to use th einductive miner is `discover_inductive()` (without the "-y") rather than `discovery_inductive`, which is what it says in the README.md.

I note that the function for the alpha miner is `discovery_alpha`, with a "-y". Should this be without the "-y"?